requirements.txt
python-dateutil==2.8.0
pip==19.2.3
torch==1.8.1
torchvision==0.9.1
pillow==6.1
dgl-cuda11.1==0.6.1
numpy==1.16.4
matplotlib==3.1.0
tensorboard==1.14.0
tensorboardx==1.8
future==0.18.2
absl-py
networkx==2.3
scikit-learn==0.21.2
scipy==1.3.0
notebook==6.0.0
h5py==2.9.0
ipykernel==5.1.2
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
plotly==4.1.1
scikit-image==0.15.0
requests==2.22.0
tqdm==4.43.0
tensorflow-gpu==2.1.0
tensorflow-estimator==2.1.0
tensorboard==2.1.1